Render cards by function name.
This will allow us to create other kinds of cards, such as a D20 dice
card more easily, since a card is just a function that takes an empty
div element and renders itself inside it.

This may also come in handy should we choose to do transitions between
cards. We can render in a div offscreen and slide it in, as an example.