Makes the 'next card' button more obvious.
I get rid of the 'ghost' css styles and use the default button styles instead. The ghost style causes confusion because it is not obvious that the arrow is a button.

I also add a message to say 'next card' and move the button to the bottom. This way the button is less likely to overlap the pixel art in portrait and landscape mode.