Adds headless Jasmine tests to Who Goes First.
See http://simonsmith.io/headless-testing-with-phantomjs-and-grunt/

I only add a simple test as an example. Now that the framework is in
place we can start creating more tests.

Also, to make the global namespace less polluted, I started using the
wgf object as a simple namespace to keep things more organized.