Deck now shuffles the first time through.
There was a bug where the cards always appeared in alphabetical order
the first time through. Now they appear shuffled.

Also, I modified the shuffling to always keep the title card as the
first card. That way we know when we have gone through the whole deck.
Also, that is the only card which does not tell you how to pick the
first player (it only has instructions on how to pick a new card). If
it is going to appear again, appearing as the first card in a freshly
shuffled deck is the only place it makes sense.