Makes card deck initialization more explicit.
This should just be a refactoring. It pulls some global state out and
puts it into a Deck constructor. (Which is not quite a JavaScript
object constructor, since I find that the closure style used by D3.js
is a little clearer)

This will make it easier to load a saved deck state from LocalStorage
in main.js instead of hiding initialization code in card.js.