Remember the last card the player was on.
Every time the user goes to the next card, we save the current state
(asynchronously, using the chrome storage APIs).

This means that if the user only uses the app once a week for game
night, they’ll be guaranteed to see a new card until they make it all
the way through the deck. (This is how I use the app.)