Fixes bug with deck loading for first time app is run.
If there is no saved state, then the app would crash on startup because
we try to access deck.cardKeys and deck.currentCard of an undefined
object.

I also add some TODOs for what I need to do to fix loading the next
time I add or remove a card.