Add description text below the next card button.
This will make the app easier to learn how to use.
http://edwardsanchez.me/blog/13589712 Even though it might end up being
more cluttered, I want people not to be frustrated when they first open
the app. It may make sense to take Edward’s advice and hide the labels
after we are confident the player knows how to use the app.