Handle new and removed cards correctly.
Deleted cards should be removed from the saved deck. New cards should be
added to the end and shuffled in to the remaining cards.

To make this easier to test, I changed loadDeck to return the deck
configuration rather than the wgf.card.Deck object. This allows me to
check for the desired state without changing Deck to expose this
internal information.