Change Flask index to use a base 'card' template.
We will be able to reuse this template as we convert the other cards to
the flask format. We'll have to think about how best to do localization,
but it would be nice to have the English text right in the source file,
and also be able to use HTML to format links and such.