Add hreflang tags and translation links.
The hreflang tags let Google know this page is translated.
https://support.google.com/webmasters/answer/189077?hl=en

The links are a simple menu to switch to the translated version of the
page.

Some clean-up: import the flask module instead of adding things to the global
namespace. Remove the CSS absolute positioning. This makes the images
too small, but I'd rather be able to see what I'm doing right now.