Add the "award" card to the Flask implementation.
First attempt at adding a new card to the Flask implementation.

It's a bit more typing than I'd like, plus it will have to change some
to allow the "random card" button to work. Also, the babel command seems
to want to translate everything in the .nox directory, so I had to make
sure to delete that before extracting messages.