Add 'random-card' page with a list of cards to pick from.
This fixes the broken link in the 'next card' arrow button on all
pages. The page will work with or without JavaScript. I hope to
catch the link event when JavaScript is enabled and go to the next
card in a deck stored in local storage (like how the Chrome / mobile
app works).

Also, makes a slight change to index.css to make the image size the
same on the about page as it is on the main page. The width wasn't
wide enough when there wasn't a full line of text.

Adds translations for the random card page.