Add session to freeze the app.
This is a poor-man's test. It'll make sure that there are no 404s or
server errors. When I release Frozen-Flask 0.13 it will also ensure
there are no redirects.

I would still want to add some kind of "test" that at a minimum checked
that there is content at some or all of the desired pages. Maybe some
WebDriver tests or something, too?